feat: add server join feature with auto-refresh and copy IP

+++ xmcl-electron-app/postinstall.ts
@@ -21,14 +21,21 @@ if (process.platform === 'linux' || process.platform === 'openbsd' || process.pl
   const fpmTargetFilePath = './node_modules/app-builder-lib/out/targets/FpmTarget.js'
   const linuxTargetHelperFilePath = './node_modules/app-builder-lib/out/targets/LinuxTargetHelper.js'
   const linuxAfterInstallShPath = './node_modules/app-builder-lib/templates/linux/after-install.tpl'
-  writeFileSync(fpmTargetFilePath, readFileSync(fpmTargetFilePath, 'utf-8')
-    // eslint-disable-next-line no-template-curly-in-string
-    .replace('installPrefix}/${appInfo.sanitizedProductName}', 'installPrefix}/xmcl'), 'utf-8')
-  writeFileSync(linuxTargetHelperFilePath, readFileSync(linuxTargetHelperFilePath, 'utf-8')
-    // eslint-disable-next-line no-template-curly-in-string
-    .replace('installPrefix}/${appInfo.sanitizedProductName}', 'installPrefix}/xmcl'), 'utf-8')
-  writeFileSync(linuxAfterInstallShPath, readFileSync(linuxAfterInstallShPath, 'utf-8')
-    // eslint-disable-next-line no-template-curly-in-string
-    .replaceAll('${sanitizedProductName}', 'xmcl'), 'utf-8')
+  
+  if (existsSync(fpmTargetFilePath)) {
+    writeFileSync(fpmTargetFilePath, readFileSync(fpmTargetFilePath, 'utf-8')
+      // eslint-disable-next-line no-template-curly-in-string
+      .replace('installPrefix}/${appInfo.sanitizedProductName}', 'installPrefix}/xmcl'), 'utf-8')
+  }
+  if (existsSync(linuxTargetHelperFilePath)) {
+    writeFileSync(linuxTargetHelperFilePath, readFileSync(linuxTargetHelperFilePath, 'utf-8')
+      // eslint-disable-next-line no-template-curly-in-string
+      .replace('installPrefix}/${appInfo.sanitizedProductName}', 'installPrefix}/xmcl'), 'utf-8')
+  }
+  if (existsSync(linuxAfterInstallShPath)) {
+    writeFileSync(linuxAfterInstallShPath, readFileSync(linuxAfterInstallShPath, 'utf-8')
+      // eslint-disable-next-line no-template-curly-in-string
+      .replaceAll('${sanitizedProductName}', 'xmcl'), 'utf-8')
+  }
   console.log('Patched linux build target')
 }
